Add negotiation helper regression tests
