+++ crates/protocol/src/version/tests/advertisement.rs
@@ -1,0 +1,21 @@
+use crate::error::NegotiationError;
+
+use super::ProtocolVersion;
+
+#[test]
+fn clamps_future_versions_in_peer_advertisements_directly() {
+    let negotiated = ProtocolVersion::from_peer_advertisement(40).expect("future versions clamp");
+    assert_eq!(negotiated, ProtocolVersion::NEWEST);
+}
+
+#[test]
+fn rejects_peer_advertisements_older_than_supported_range() {
+    let err = ProtocolVersion::from_peer_advertisement(27).unwrap_err();
+    assert_eq!(err, NegotiationError::UnsupportedVersion(27));
+}
+
+#[test]
+fn rejects_zero_peer_advertisement() {
+    let err = ProtocolVersion::from_peer_advertisement(0).unwrap_err();
+    assert_eq!(err, NegotiationError::UnsupportedVersion(0));
+}
